Fix database tests
Had to make Sakskompleks a data class again, because sykmeldinger wasn't saved to the database.
Also had to override equals (and hashcode) for Sakskomplekstilstand, and therefore making it an abstract class because an interface can't have equals().

There's probably more ways to solve this - Ideally, I would like to hide the lists from the users, but still be able to serialize them to the database.

There are still some tests failing, but they concern multiple sykmeldinger, which the state machine doesn't support yet. We have to decide if we should remove this functionality, or extend the state machine.